Update protected-disease-gene.tsv
I added OMIM:219700 for gene association to keep, even though it does not follow our rules. 
The "type" does not exist yet. 
Basically, in omim, the disease itself is associated to the gene in this list. The other associated genes in OMIM are for susceptibility to or modifier of the disease.